refactor: split portable DNS helpers into platform/posix/dns_common.c
